+++ cli/commands/postgres.go
@@ -5,9 +5,6 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
-
-	"github.com/wundergraph/wundergraph/pkg/datasources/database"
-	"github.com/wundergraph/wundergraph/pkg/files"
 )
 
 type DatabaseIntrospectionResult struct {
@@ -23,27 +20,12 @@ var postgresCmd = &cobra.Command{
 	Example: `wunderctl introspect postgresql postgresql://user:password@localhost:5432/database?schema=public`,
 	Args:    cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		wunderGraphDir, err := files.FindWunderGraphDir(_wunderGraphDirConfig)
-		if err != nil {
-			return err
-		}
-
 		databaseURL := args[0]
 		introspectionSchema := fmt.Sprintf(`datasource db {
 			provider = "postgresql"
 			url      = "%s"
 		}`, databaseURL)
-		prismaSchema, graphqlSDL, dmmf, err := database.IntrospectPrismaDatabase(introspectionSchema, wunderGraphDir, log)
-		if err != nil {
-			return err
-		}
-		result := DatabaseIntrospectionResult{
-			PrismaSchema:  prismaSchema,
-			GraphQLSchema: graphqlSDL,
-			Dmmf:          []byte(dmmf),
-		}
-		emitIntrospectionResult(result)
-		return nil
+		return introspectDatabase(introspectionSchema, true)
 	},
 }