+++ packages/sdk/src/client/InputValidationError.ts
@@ -1,19 +1,21 @@
-type error = {
+import { ResponseError } from './ResponseError';
+
+type ValidationError = {
 	propertyPath: string;
 	message: string;
 	invalidValue: any;
 };
 
-interface ResponseJson {
-	errors: error[];
+interface ValidationResponseJSON {
+	errors: ValidationError[];
 	message: string;
 	input: object;
 }
 
-export class InputValidationError extends Error {
-	public errors: error[];
-	constructor(public json: ResponseJson, public statusCode: number) {
-		super(json.message);
+export class InputValidationError extends ResponseError {
+	public errors: ValidationError[];
+	constructor(json: ValidationResponseJSON, public statusCode: number) {
+		super(json.message, statusCode);
 		this.name = 'InputValidationError';
 		this.errors = json.errors;
 		// because we are extending a built-in class