+++ packages/sdk/src/codegen/templates/typescript/hooks.ts
@@ -3,24 +3,25 @@ import { CodeGenerationConfig } from '../../../configure';
 import Handlebars from 'handlebars';
 import { formatTypeScript } from './index';
 import { OperationType } from '@wundergraph/protobuf';
-import { modelImports, operations } from './helpers';
+import { filterNodeJSOperations, modelImports, operations } from './helpers';
 import { template } from './hooks.template';
 import templates from '../index';
 
 export class WunderGraphHooksPlugin implements Template {
 	generate(config: CodeGenerationConfig): Promise<TemplateOutputFile[]> {
+		const application = filterNodeJSOperations(config.application);
 		const tmpl = Handlebars.compile(template);
-		const _queries = operations(config.application, OperationType.QUERY, false);
-		const _internalQueries = operations(config.application, OperationType.QUERY, true);
-		const _mutations = operations(config.application, OperationType.MUTATION, false);
-		const _internalMutations = operations(config.application, OperationType.MUTATION, true);
-		const _subscriptions = operations(config.application, OperationType.SUBSCRIPTION, false);
-		const _uploadProviders = config.application.S3UploadProvider;
+		const _queries = operations(application, OperationType.QUERY, false);
+		const _internalQueries = operations(application, OperationType.QUERY, true);
+		const _mutations = operations(application, OperationType.MUTATION, false);
+		const _internalMutations = operations(application, OperationType.MUTATION, true);
+		const _subscriptions = operations(application, OperationType.SUBSCRIPTION, false);
+		const _uploadProviders = application.S3UploadProvider;
 		const content = tmpl({
-			modelImports: modelImports(config.application, true),
-			operationNamesUnion: config.application.Operations.map((o) => `"${o.Name}"`).join(' | ') || 'never',
+			modelImports: modelImports(application, true),
+			operationNamesUnion: application.Operations.map((o) => `"${o.Name}"`).join(' | ') || 'never',
 			dataSourcesUnion:
-				config.application.EngineConfiguration.DataSources.filter((ds) => ds.Id !== undefined && ds.Id !== '')
+				application.EngineConfiguration.DataSources.filter((ds) => ds.Id !== undefined && ds.Id !== '')
 					.map((ds) => `"${ds.Id}"`)
 					.filter((ds, i, arr) => arr.indexOf(ds) === i)
 					.join(' | ') || 'never',