+++ packages/protobuf/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.103.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/protobuf@0.102.0...@wundergraph/protobuf@0.103.0) (2023-01-30)
+
+### Features
+
+* add prisma datasource ([#605](https://github.com/wundergraph/wundergraph/issues/605)) ([2f0cfa5](https://github.com/wundergraph/wundergraph/commit/2f0cfa5e420080772a6552f4a256c7d71906f84a)) (@jensneuse)
+
 ## [0.102.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/protobuf@0.101.0...@wundergraph/protobuf@0.102.0) (2023-01-20)
 
 ### Features